Documentation and renaming
Another round of polishing and documentation for the Discord Bot List API wrapper.
