+++ DiscordBotsList.Api/Internal/SelfBot.cs
@@ -7,21 +7,21 @@ namespace DiscordBotsList.Api.Internal
 	internal class SelfBot : Bot, IDblSelfBot
 	{
 		public async Task<List<IDblEntity>> GetVotersAsync()
-			=> await ((AuthDiscordBotListApi) api).GetVotersAsync();
+			=> await ((DblClient) api).GetVotersAsync();
 
 	    public async Task<bool> HasVotedAsync(ulong userId)
-	        => await ((AuthDiscordBotListApi) api).HasVoted(userId);
+	        => await ((DblClient) api).HasVotedAsync(userId);
 
 	    public async Task<bool> IsWeekendAsync()
-	        => await ((AuthDiscordBotListApi) api).IsWeekendAsync();
+	        => await ((DblClient) api).IsWeekendAsync();
 
 		public async Task UpdateStatsAsync(int guildCount)
-			=> await ((AuthDiscordBotListApi) api).UpdateStats(guildCount);
+			=> await ((DblClient) api).UpdateStatsAsync(guildCount);
 
 		public async Task UpdateStatsAsync(int[] shards)
-			=> await ((AuthDiscordBotListApi) api).UpdateStats(0, shards.Length, shards);
+			=> await ((DblClient) api).UpdateStatsAsync(0, shards.Length, shards);
 
 		public async Task UpdateStatsAsync(int shardCount, int totalShards, params int[] shards)
-			=> await ((AuthDiscordBotListApi) api).UpdateStats(shardCount, totalShards, shards);
+			=> await ((DblClient) api).UpdateStatsAsync(shardCount, totalShards, shards);
 	}
 }
