+++ DiscordBotsList.Api/Objects/DblStatus.cs
@@ -1,0 +1,11 @@
+﻿namespace DiscordBotsList.Api.Objects
+{
+    public enum DblStatus
+	{
+		Supporter,
+		Certified,
+		Moderator,
+		WebModerator,
+		Admin
+	}
+}
