+++ DiscordBotsList.Api/Objects/IAdapter.cs
@@ -3,6 +3,9 @@ using System.Threading.Tasks;
 
 namespace DiscordBotsList.Api.Objects
 {
+	/// <summary>
+	/// Represents a generic adapter.
+	/// </summary>
     public interface IAdapter
     {
 		event Action<string> Log;
