+++ DiscordBotsList.Api/Objects/IDblEntity.cs
@@ -1,24 +1,27 @@
 ﻿namespace DiscordBotsList.Api.Objects
 {
-    public interface IDblEntity
+	/// <summary>
+	/// Represents a generic entity for the Discord Bot List API.
+	/// </summary>
+	public interface IDblEntity
 	{
 		/// <summary>
-		/// Discord Id
+		/// Represents the ID of this <see cref="IDblEntity"/>.
 		/// </summary>
 		ulong Id { get; }
 
 		/// <summary>
-		/// Username of the entity
+		/// Represents the username of this <see cref="IDblEntity"/>.
 		/// </summary>
 		string Username { get; }
 
 		/// <summary>
-		/// Discriminator, the XXXX#1234 part
+		/// Represents the discriminator for this <see cref="IDblEntity"/>, which is located after the username.
 		/// </summary>
 		string Discriminator { get; }
 		
 		/// <summary>
-		/// Discord avatar url, or default avatar if none found.
+		/// Represents the avatar URL for this <see cref="IDblEntity"/>, or a default avatar URL if none could be found.
 		/// </summary>
 		string AvatarUrl { get; }
 	}
