+++ DiscordBotsList.Api/Objects/SmallWidgetType.cs
@@ -1,0 +1,11 @@
+﻿namespace DiscordBotsList.Api.Objects
+{
+    public enum SmallWidgetType
+	{
+        STATUS,
+        SERVERS,
+        LIB,
+        UPVOTES,
+        OWNER
+    }
+}
