+++ DiscordBotsList.Api/Objects/WeekendObject.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace DiscordBotsList.Api.Objects
+﻿namespace DiscordBotsList.Api.Objects
 {
     public struct WeekendObject
     {
