+++ DiscordBotsList.Api/Objects/WidgetImageFormat.cs
@@ -1,0 +1,11 @@
+﻿namespace DiscordBotsList.Api.Objects
+{
+    /// <summary>
+    /// Defines the image formats that can be set for a widget.
+    /// </summary>
+    public enum WidgetImageFormat
+    {
+        Svg = 1,
+        Png = 2
+    }
+}
